Issue #11434 - resolve differences in AliasCheckerSymlinkTests
