+++ jetty-ee10/jetty-ee10-servlet/src/main/java/org/eclipse/jetty/ee10/servlet/ServletContextHandler.java
@@ -1197,7 +1197,7 @@ public class ServletContextHandler extends ContextHandler
             return true;
         }
 
-        return super.handleByContextHandler(pathInContext, request, response, callback);
+        return false;
     }
 
     @Override
