--- jetty-ee10/jetty-ee10-tests/jetty-ee10-test-integration/src/test/resources/altDir1/file1
@@ -1,1 +1,0 @@
-file 1 contents
