--- jetty-ee10/jetty-ee10-tests/jetty-ee10-test-integration/src/test/resources/altDir2/file2
@@ -1,1 +1,0 @@
-file 2 contents
