--- jetty-ee10/jetty-ee10-tests/jetty-ee10-test-integration/src/test/resources/sibling/file
@@ -1,1 +1,0 @@
-This file is inside a sibling dir to webroot.
