--- jetty-ee10/jetty-ee10-tests/jetty-ee10-test-integration/src/test/resources/webroot/WEB-INF/web.xml
@@ -1,1 +1,0 @@
-This is the web.xml file.
