--- jetty-ee10/jetty-ee10-tests/jetty-ee10-test-integration/src/test/resources/webroot/documents/file
@@ -1,1 +1,0 @@
-This file is inside webroot/documents.
