--- jetty-ee10/jetty-ee10-tests/jetty-ee10-test-integration/src/test/resources/webroot/index.html
@@ -1,4 +1,0 @@
-<html>
-<h1>hello world</h1>
-<p>body of index.html</p>
-</html>
