+++ jetty-ee9/jetty-ee9-tests/jetty-ee9-test-integration/src/test/resources/combined/WEB-INF/file
@@ -1,0 +1,1 @@
+This is in the second WEB-INF dir.
