+++ jetty-ee9/jetty-ee9-tests/jetty-ee9-test-integration/src/test/resources/combined/combinedFile
@@ -1,0 +1,1 @@
+This is a file in the combined resource dir.
