+++ jetty-ee9/jetty-ee9-tests/jetty-ee9-test-integration/src/test/resources/combined/files/file1
@@ -1,0 +1,1 @@
+file1 from combined dir
