+++ tests/test-integration/src/test/resources/combined/WEB-INF/file
