+++ tests/test-integration/src/test/resources/combined/combinedFile
