+++ tests/test-integration/src/test/resources/combined/files/file1
